move generate_resources
