+++ test/kubernetes/e2e/features/headless_svc/gloo_gateway_suite.go
@@ -8,7 +8,7 @@ import (
 	"github.com/solo-io/gloo/pkg/utils/requestutils/curl"
 	"github.com/solo-io/gloo/projects/gateway/pkg/defaults"
 	"github.com/solo-io/gloo/test/kubernetes/e2e"
-	"github.com/solo-io/gloo/test/kubernetes/e2e/utils"
+	"github.com/solo-io/gloo/test/kubernetes/testutils/resources"
 	"github.com/stretchr/testify/suite"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
@@ -37,8 +37,8 @@ func NewEdgeGatewayHeadlessSvcSuite(ctx context.Context, testInst *e2e.TestInsta
 
 // SetupSuite generates manifest files for the test suite
 func (s *edgeGatewaySuite) SetupSuite() {
-	resources := GetEdgeGatewayResources(s.testInstallation.Metadata.InstallNamespace)
-	err := utils.WriteResourcesToFile(resources, s.routingManifestFile)
+	gwResources := GetEdgeGatewayResources(s.testInstallation.Metadata.InstallNamespace)
+	err := resources.WriteResourcesToFile(gwResources, s.routingManifestFile)
 	s.Require().NoError(err, "can write resources to file")
 }
