+++ test/kubernetes/testutils/resources/generate_resources.go
@@ -1,4 +1,4 @@
-package utils
+package resources
 
 import (
 	"bufio"
